Header, Settings and Profile pages done
The DropDown Menu in the Header now properly displays the users icon, name and email. and it gives the options to go to their profile or the settings menu and also an option to logout.

settings page allows the user to change their first, middle and last name. and also the location